requirements.txt
# Web Cartographer (Website Explorer Agent)
strands-agents[otel]>=1.0.0
strands-agents-tools>=0.2.0
playwright>=1.40.0
neo4j>=5.0.0
Pillow>=10.0.0
aiohttp>=3.9.0

# Release Revert Risk Advisor Agent
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
requests==2.32.3
pyyaml==6.0.2
datadog-api-client==2.31.0

# Shared dependencies
boto3>=1.36.4
python-dotenv>=1.0.1
opentelemetry-exporter-otlp-proto-http>=1.20.0
